changed IMathTree into an absract class and parsed implementations into their own .cs files. added readme

+++ Calculator.Tests/CalcTests.cs
@@ -7,13 +7,13 @@ namespace CommandLineCalc.Tests
     {
 
         [DataTestMethod]
-        [DataRow("3 + 4",7)]
+        [DataRow("3 + 4", 7)]
         [DataRow("3 + 4 + 3", 10)]
         [DataRow("3.5 + 6.5 + 2.0 + 1", 13)]
         public void AdditionTests(string input, double solution)
         {
-            double sol = Calculator.Calculate(input);
-            Assert.AreEqual(solution, sol);
+            decimal sol = Calculator.Calculate(input);
+            Assert.AreEqual((decimal)solution, sol);
         }
 
 
@@ -26,8 +26,8 @@ namespace CommandLineCalc.Tests
         [DataRow("10 - 3 + 4", 11)]
         public void SubtractionTests(string input, double solution)
         {
-            double sol = Calculator.Calculate(input);
-            Assert.AreEqual(solution, sol);
+            decimal sol = Calculator.Calculate(input);
+            Assert.AreEqual((decimal)solution, sol);
         }
 
 
@@ -40,8 +40,8 @@ namespace CommandLineCalc.Tests
         [DataRow("4 * 5 - 2 * 3", 14)]
         public void MultiplicationTests(string input, double solution)
         {
-            double sol = Calculator.Calculate(input);
-            Assert.AreEqual(solution, sol);
+            decimal sol = Calculator.Calculate(input);
+            Assert.AreEqual((decimal)solution, sol);
         }
 
         
@@ -51,8 +51,8 @@ namespace CommandLineCalc.Tests
         [DataRow("16 / 2 / 2", 4)]
         public void DivisionTests(string input, double solution)
         {
-            double sol = Calculator.Calculate(input);
-            Assert.AreEqual(solution, sol);
+            decimal sol = Calculator.Calculate(input);
+            Assert.AreEqual((decimal)solution, sol);
         }
 
 
@@ -62,8 +62,8 @@ namespace CommandLineCalc.Tests
         [DataRow("3 ^ 4 ^ 5", 3486784401)]
         public void ExponentTests(string input, double solution)
         {
-            double sol = Calculator.Calculate(input);
-            Assert.AreEqual(solution, sol);
+            decimal sol = Calculator.Calculate(input);
+            Assert.AreEqual((decimal)solution, sol);
         }
 
 
@@ -87,8 +87,8 @@ namespace CommandLineCalc.Tests
         [DataRow("(3 + 2) ^ (1 + 1)", 25)]
         public void ParenthesisTests(string input, double solution)
         {
-            double sol = Calculator.Calculate(input);
-            Assert.AreEqual(solution, sol);
+            decimal sol = Calculator.Calculate(input);
+            Assert.AreEqual((decimal)solution, sol);
         }
     }
 }
